+++ packages/web-workers/src/messenger/iframe/constants.ts
@@ -1,0 +1,1 @@
+export const IFRAME_RUN_IDENTIFIER = '__run';
