+++ packages/web-workers/src/messenger/iframe/remote-origin.ts
@@ -1,0 +1,9 @@
+import {createIframeWorkerMessenger} from './create-messenger';
+
+export function createRemoteOriginIframeMessengerFactory(iframeSrc: string) {
+  return (url: URL) => {
+    return createIframeWorkerMessenger(url, (iframe) => {
+      iframe.src = iframeSrc;
+    });
+  };
+}
